requirements.txt
langchain>=1.0.0,<1.2
langgraph>=1.0.0,<1.2
langchain-openai>=1.1.0,<1.3
langchain-community>=0.3.0,<0.5
langchain-text-splitters>=0.3.0,<0.4
weaviate-client>=3.25.3,<4.0
dashscope>=1.20.0,<2.0
python-dotenv>=1.0.0,<1.1
streamlit>=1.32.0,<2.0
